Bug Patch - Consistency in font names
Font names in tagsedit and the note editor aren't the same when they
aren't saved in the tags.xml.
In tagsedit, that's the index(0) which is used (ie the first font
alphabetically) and in the note editor the correct default font the
system.
This is somewhat misleading => patch to use the default font in tagsedit
too.