Tests for a template system
Very very minimal and simplistic template system for a start :
it only consists of an xml file identic to those which are copied in the
basket folder by the basketFactory class.
The unpack consists then onlyto a QFile::Copy call (when I said minimal
and simplistic :-) ).

I've added an option from BNPView to generate these files from an actual
template, from the menu or the popup.

The template can be used from new template dialog :
the icon, background image, font... are reused.
The dialog icon and color chooser reflect the choce when a template is
clicked.

Note : I've added a lot of kDebugs() calsl  I will delete.