Tests template - some work on newbasketdialog
Just done some work on the newbasketdialog form to reflect the
template properties on the icon and the control when selecting.
The code is a bit ugly, the template reading needs to be put
in a distinct class (a lot of redundancy rght now).
It's a very preliminary work to see what needs to be done for
a good template system.