Templates - Management dialog
Started to work on managing the templates.
Right now, copy and delete are done.
Renamed will be done shortly, acces to basket properties later.
I've created a signal to update the NewBasketDialog dialog.

It's received but not used to refresh the WidgetList
right now.

I've  created an input dialog for the template name, so the actual
code fo crreating a template from a basket will be reworked.
(elminite redundancy).

Committer: Olivier Rollet <o.rollet@gmail.com>
	modified:   src/.directory
	modified:   src/CMakeLists.txt
	new file:   src/managetemplatedialog.cpp
	new file:   src/managetemplatedialog.h
	new file:   src/namedialog.cpp
	new file:   src/namedialog.h
	modified:   src/newbasketdialog.cpp
	modified:   src/newbasketdialog.h
	modified:   src/templatefactory.cpp
	modified:   src/templatefactory.h
	modified:   src/totemplatedialog.cpp